chore(utils): rename e2e/helpers to e2e/utils
