+++ tests/e2e/runner.py
@@ -9,15 +9,14 @@ from bom_open import bom_open
 
 from click import unstyle
 
-from helpers.StoryscriptHubFixture import StoryscriptHubFixture
-
 from pytest import fixture, mark
 
 from storyscript.Api import Api
 import storyscript.Hub as StoryHub
 from storyscript.App import _clean_dict
 
-from tests.e2e.helpers.Features import parse_features
+from tests.e2e.utils.StoryscriptHubFixture import StoryscriptHubFixture
+from tests.e2e.utils.Features import parse_features
 
 test_dir = path.dirname(path.realpath(__file__))
 # make the test_file paths relative, s.t. test paths are nice to read
