+++ tests/fixtures/update_services_fixtures
@@ -7,7 +7,7 @@ import sys
 sys.path.insert(
     0, path.dirname(path.dirname(path.dirname(path.realpath(__file__)))))
 
-from tests.e2e.helpers.StoryscriptHubFixture import StoryscriptHubFixture
+from tests.e2e.utils.StoryscriptHubFixture import StoryscriptHubFixture
 
 
 services = [
