+++ tests/unittests/StoryscriptHubFixture.py
@@ -5,7 +5,7 @@ from os import path
 sys.path.insert(
     0, path.dirname(path.dirname(path.dirname(path.realpath(__file__)))))
 
-from tests.e2e.helpers.StoryscriptHubFixture import ServiceWrapper, \
+from tests.e2e.utils.StoryscriptHubFixture import ServiceWrapper, \
     StoryscriptHubFixture   # noqa: E402
